Prevent blockchain corruption by terminating upon bitcoin.pid

+++ run.sh
@@ -1,28 +1,33 @@
-if [ ! -d "src" ]; then
+if [ ! -d "src" ] ; then
 	cd .. # If located in src or any other level of depth 1, return
 fi
 
 pruned="false"
-if [ -d "/media/sf_Bitcoin/blocks/" ]; then
+if [ -d "/media/sf_Bitcoin/blocks/" ] && [ ! -f "/media/sf_Bitcoin/bitcoin.pid" ] ; then
 	dir="/media/sf_Bitcoin"
-elif [ -d "/media/sf_BitcoinAttacker/blocks/" ]; then
+elif [ -d "/media/sf_BitcoinAttacker/blocks/" ] && [ ! -f "/media/sf_BitcoinAttacker/bitcoin.pid" ] ; then
 	dir="/media/sf_BitcoinAttacker"
-elif [ -d "/media/sf_BitcoinVictim/blocks/" ]; then
+elif [ -d "/media/sf_BitcoinVictim/blocks/" ] && [ ! -f "/media/sf_BitcoinVictim/bitcoin.pid" ] ; then
 	dir="/media/sf_BitcoinVictim"
-elif [ -d "/media/sim/BITCOIN/blocks/" ]; then
+elif [ -d "/media/sim/BITCOIN/blocks/" ] && [ ! -f "/media/sim/BITCOIN/bitcoin.pid" ] ; then
 	dir="/media/sim/BITCOIN"
 else
 	dir="$HOME/.bitcoin/"
 	pruned="true"
 
-	if [ ! -d "$dir" ]; then
+	if [ ! -d "$dir" ] ; then
 		mkdir "$dir"
 	fi
 fi
 
+if [ -f "$dir/bitcoin.pid" ] ; then
+	echo "The directory \"$dir\" has bitcoin.pid, meaning that Bitcoin is already running. In order to ensure that the blockchain does not get corrupted, the program will now terminate."
+	exit 1
+fi
+
 echo "datadir = $dir"
 
-if [ ! -f "$dir/bitcoin.conf" ]; then
+if [ ! -f "$dir/bitcoin.conf" ] ; then
 	echo "Resetting configuration file"
 	echo "server=1" > "$dir/bitcoin.conf"
 	echo "rpcuser=cybersec" >> "$dir/bitcoin.conf"
@@ -34,8 +39,8 @@ if [ ! -f "$dir/bitcoin.conf" ]; then
 	echo "listen=1" >> "$dir/bitcoin.conf"
 fi
 
-if [ "$1" == "gui" ]; then
-	if [ "$pruned" == "true" ]; then
+if [ "$1" == "gui" ] ; then
+	if [ "$pruned" == "true" ] ; then
 		echo "Pruned mode activated, only keeping 550 block transactions"
 		echo
 		src/qt/bitcoin-qt -prune=550 -datadir="$dir" -debug=researcher
@@ -46,23 +51,25 @@ if [ "$1" == "gui" ]; then
 else
 
 	# Only open the console if not already open
-	if ! wmctrl -l | grep -q "Custom Bitcoin Console"; then
+	if ! wmctrl -l | grep -q "Custom Bitcoin Console" ; then
 		# Find the right terminal
-		if [ -x "$(command -v mate-terminal)" ]; then
+		if [ -x "$(command -v mate-terminal)" ] ; then
 			mate-terminal -t "Custom Bitcoin Console" -- python3 bitcoin_console.py
-		elif [ -x "$(command -v xfce4-terminal)" ]; then
+		elif [ -x "$(command -v xfce4-terminal)" ] ; then
 			xfce4-terminal -t "Custom Bitcoin Console" -- python3 bitcoin_console.py
 		else
 			gnome-terminal -t "Custom Bitcoin Console" -- python3 bitcoin_console.py
 		fi
 	fi
 
-	if [ "$pruned" == "true" ]; then
+	if [ "$pruned" == "true" ] ; then
 		echo "Pruned mode activated, only keeping 550 block transactions"
 		echo
 		src/bitcoind -prune=550 -datadir="$dir" -debug=researcher
 	else
 		echo
 		src/bitcoind -datadir="$dir" -debug=researcher
+		# Reindexing the chainstate:
+		#src/bitcoind -datadir="/media/sf_Bitcoin" -debug=researcher -reindex-chainstate
 	fi
 fi
